moved internal utils to utils project » github.com/RunedUniverse/utils
