+++ rogm-core/src/main/java/net/runeduniverse/libs/rogm/errors/ScannerException.java
@@ -1,4 +1,6 @@
-package net.runeduniverse.libs.rogm.error;
+package net.runeduniverse.libs.rogm.errors;
+
+import net.runeduniverse.libs.errors.ExceptionSuppressions;
 
 public class ScannerException extends ExceptionSuppressions {
 	private static final long serialVersionUID = 4295891405650593671L;
