+++ rogm-core/src/main/java/net/runeduniverse/libs/rogm/modules/PassiveModule.java
@@ -1,6 +1,6 @@
 package net.runeduniverse.libs.rogm.modules;
 
-import net.runeduniverse.libs.rogm.error.ScannerException;
+import net.runeduniverse.libs.rogm.errors.ScannerException;
 import net.runeduniverse.libs.rogm.pattern.Archive;
 
 public interface PassiveModule {
