+++ rogm-core/src/main/java/net/runeduniverse/libs/rogm/pattern/Archive.java
@@ -13,9 +13,9 @@ import lombok.Getter;
 import net.runeduniverse.libs.logging.logs.CompoundTree;
 import net.runeduniverse.libs.rogm.annotations.IConverter;
 import net.runeduniverse.libs.rogm.annotations.Id;
-import net.runeduniverse.libs.rogm.error.ScannerException;
+import net.runeduniverse.libs.rogm.errors.ScannerException;
 import net.runeduniverse.libs.rogm.info.PackageInfo;
-import net.runeduniverse.libs.rogm.logging.Level;
+import net.runeduniverse.libs.logging.Level;
 import net.runeduniverse.libs.rogm.modules.IdTypeResolver;
 import net.runeduniverse.libs.rogm.pattern.IPattern.IDeleteContainer;
 import net.runeduniverse.libs.rogm.pattern.scanner.TypeScanner;
