+++ rogm-core/src/main/java/net/runeduniverse/libs/rogm/pipeline/chain/AssemblyLayers.java
@@ -7,6 +7,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import net.runeduniverse.libs.chain.Chain;
+import net.runeduniverse.libs.chain.ChainRuntime;
 import net.runeduniverse.libs.rogm.annotations.Direction;
 import net.runeduniverse.libs.rogm.annotations.PostLoad;
 import net.runeduniverse.libs.rogm.annotations.PostReload;
@@ -22,8 +24,6 @@ import net.runeduniverse.libs.rogm.pattern.IPattern.IDataRecord;
 import net.runeduniverse.libs.rogm.pattern.IPattern.IPatternContainer;
 import net.runeduniverse.libs.rogm.pipeline.chain.data.EntityContainer;
 import net.runeduniverse.libs.rogm.pipeline.chain.data.RelatedEntriesContainer;
-import net.runeduniverse.libs.rogm.pipeline.chain.sys.Chain;
-import net.runeduniverse.libs.rogm.pipeline.chain.sys.ChainRuntime;
 import net.runeduniverse.libs.rogm.querying.IFNode;
 import net.runeduniverse.libs.rogm.querying.IFRelation;
 import net.runeduniverse.libs.rogm.querying.IFilter;
