--- rogm-core/src/main/java/net/runeduniverse/libs/rogm/pipeline/chain/sys/Chains.java
@@ -1,13 +1,0 @@
-package net.runeduniverse.libs.rogm.pipeline.chain.sys;
-
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-@Retention(RUNTIME)
-@Target(METHOD)
-public @interface Chains {
-	Chain[] value();
-}
