--- rogm-core/src/main/java/net/runeduniverse/libs/rogm/pipeline/chain/sys/ILayer.java
@@ -1,5 +1,0 @@
-package net.runeduniverse.libs.rogm.pipeline.chain.sys;
-
-public interface ILayer {
-	void call(ChainRuntime<?> runtime) throws Exception;
-}
