--- rogm-core/src/main/java/net/runeduniverse/libs/rogm/pipeline/chain/sys/Level.java
@@ -1,12 +1,0 @@
-package net.runeduniverse.libs.rogm.pipeline.chain.sys;
-
-public class Level extends java.util.logging.Level {
-	private static final long serialVersionUID = -1369172045174361996L;
-
-	protected Level(String name, int value) {
-		super(name, value);
-	}
-
-	public static final Level BURY = new Level("BURY", 601);
-
-}
