+++ rogm-lang-cypher/src/main/java/net/runeduniverse/libs/rogm/lang/cypher/CypherLanguage.java
@@ -5,10 +5,10 @@ import java.util.logging.Logger;
 import net.runeduniverse.libs.rogm.lang.Language;
 import net.runeduniverse.libs.rogm.lang.cypher.pipeline.chains.CleanupLayers;
 import net.runeduniverse.libs.rogm.lang.cypher.pipeline.chains.CypherChains;
-import net.runeduniverse.libs.rogm.logging.UniversalLogger;
+import net.runeduniverse.libs.logging.UniversalLogger;
 import net.runeduniverse.libs.rogm.modules.IdTypeResolver;
 import net.runeduniverse.libs.rogm.parser.Parser;
-import net.runeduniverse.libs.rogm.pipeline.chain.sys.ChainManager;
+import net.runeduniverse.libs.chain.ChainManager;
 
 public class CypherLanguage implements Language {
