+++ rogm-lang-cypher/src/main/java/net/runeduniverse/libs/rogm/lang/cypher/pipeline/chains/CleanupLayers.java
@@ -8,8 +8,8 @@ import net.runeduniverse.libs.rogm.lang.cypher.CypherInstance;
 import net.runeduniverse.libs.rogm.lang.cypher.FilterStatus;
 import net.runeduniverse.libs.rogm.lang.cypher.Mapper;
 import net.runeduniverse.libs.rogm.modules.Module;
-import net.runeduniverse.libs.rogm.pipeline.chain.sys.Chain;
-import net.runeduniverse.libs.rogm.pipeline.chain.sys.ChainRuntime;
+import net.runeduniverse.libs.chain.Chain;
+import net.runeduniverse.libs.chain.ChainRuntime;
 import net.runeduniverse.libs.rogm.querying.IFNode;
 import net.runeduniverse.libs.rogm.querying.IFRelation;
 import net.runeduniverse.libs.rogm.querying.IFilter;
