+++ rogm-module-neo4j/src/test/java/net/runeduniverse/libs/rogm/modules/neo4j/SessionTest.java
@@ -10,7 +10,7 @@ import java.util.logging.Logger;
 import org.junit.jupiter.api.*;
 import net.runeduniverse.libs.rogm.Configuration;
 import net.runeduniverse.libs.rogm.Session;
-import net.runeduniverse.libs.rogm.logging.DebugLogger;
+import net.runeduniverse.libs.logging.DebugLogger;
 import net.runeduniverse.libs.rogm.modules.neo4j.SessionTest;
 import net.runeduniverse.libs.rogm.pipeline.Pipeline;
 import net.runeduniverse.libs.rogm.querying.IFNode;
